Multi tenant prepared.

+++ BE/Artin.BringAuto/GraphQL/Users/UserMutation.cs
@@ -1,6 +1,7 @@
 ﻿using Artin.BringAuto.DAL;
 using Artin.BringAuto.DAL.Models;
 using Artin.BringAuto.Shared;
+using Artin.BringAuto.Shared.Ifaces;
 using Artin.BringAuto.Shared.Users;
 using AutoMapper;
 using HotChocolate.AspNetCore.Authorization;
@@ -19,42 +20,79 @@ namespace Artin.BringAuto.GraphQL.Users
         private readonly RoleManager<IdentityRole> roleManager;
         private readonly UserManager<ApplicationUser> userManager;
         private readonly IMapper mapper;
+        private readonly ICurrentUserId currentUser;
         private readonly BringAutoDbContext dbContext;
+        private readonly ICurrentTenant currentTenant;
 
         public UserMutation(RoleManager<IdentityRole> roleManager, UserManager<ApplicationUser> userManager, IMapper mapper,
-            BringAutoDbContext dbContext)
+            ICurrentUserId currentUser,
+            BringAutoDbContext dbContext,
+            ICurrentTenant currentTenant)
         {
             this.roleManager = roleManager;
             this.userManager = userManager;
             this.mapper = mapper;
+            this.currentUser = currentUser;
             this.dbContext = dbContext;
+            this.currentTenant = currentTenant;
         }
 
         [Authorize(Roles = new[] { RoleNames.Admin })]
         public async Task<IdentityResult> Add(NewUser user)
         {
+
+            if (user.Roles?.Any() != true && string.IsNullOrEmpty(user.NewTenantName))
+                return IdentityResult.Failed(new IdentityError() { Code = "Unkown role", Description = "New tenant name od roles has to be defined" });
+
             var appUser = mapper.Map<ApplicationUser>(user);
+
             var result = await userManager.CreateAsync(appUser, user.Password);
             if (result.Succeeded)
             {
 
-                if (!await dbContext.Tenants.AnyAsync(x => x.Id == user.Tenant))
+                var tenantId = currentTenant.GetTenantId();
+                if (!tenantId.HasValue)
                 {
                     await userManager.DeleteAsync(appUser);
                     return IdentityResult.Failed(new IdentityError() { Code = "Unknown tenant", Description = "Specified tenant doesn't exist" });
                 }
 
+                if (!String.IsNullOrWhiteSpace(user.NewTenantName))
+                {
+                    try
+                    {
+                        tenantId = await CreateNewTenant(user.NewTenantName);
+                    }
+                    catch (Exception ex)
+                    {
+                        await userManager.DeleteAsync(appUser);
+                        return IdentityResult.Failed(new IdentityError() { Code = "Tenant already exists", Description = ex.Message });
+                    }
+                    user.Roles ??= new List<string>();
+                    user.Roles.Add(RoleNames.Admin);
+                }
+
+
                 dbContext.Add(new UserTenancy()
                 {
                     UserId = appUser.Id,
-                    TenantId = user.Tenant
+                    TenantId = tenantId.Value
                 });
-
-                await dbContext.SaveChangesAsync();
+                try
+                {
+                    await dbContext.SaveChangesAsync();
+                }
+                catch (Exception ex)
+                {
+                    await userManager.DeleteAsync(appUser);
+                    return IdentityResult.Failed(new IdentityError() { Code = "Unknown tenant", Description = "Specified tenant doesn't exist" });
+                }
 
                 var identityResult = await userManager.AddToRolesAsync(appUser, user.Roles);
                 if (!identityResult.Succeeded)
                 {
+                    await dbContext.UserTenancy.Where(x => x.UserId == appUser.Id).ForEachAsync(x => dbContext.Remove(x));
+                    await dbContext.SaveChangesAsync();
                     await userManager.DeleteAsync(appUser);
                     return identityResult;
                 }
@@ -64,8 +102,42 @@ namespace Artin.BringAuto.GraphQL.Users
             return result;
         }
 
+        private async Task<int> CreateNewTenant(string name)
+        {
+            var tenant = await dbContext.Tenants.FirstOrDefaultAsync(x => x.Name == name);
+
+            if (tenant is null)
+            {
+                tenant = new Tenant()
+                {
+                    Name = name
+                };
+                dbContext.Tenants.Add(tenant);
+            }
+            else
+            {
+                if (!await dbContext.UserTenancy.AnyAsync(x => x.UserId == currentUser.CurrentUserId && x.TenantId == tenant.Id))
+                {
+                    throw new ArgumentException("Current admin is not part of specified tenant");
+                }
+            }
+            await dbContext.SaveChangesAsync();
+
+            if (!await dbContext.UserTenancy.AnyAsync(x => x.UserId == currentUser.CurrentUserId && x.TenantId == tenant.Id))
+            {
+                var userTenancy = new UserTenancy()
+                {
+                    Tenant = tenant,
+                    UserId = currentUser.CurrentUserId
+                };
+                dbContext.UserTenancy.Add(userTenancy);
+                await dbContext.SaveChangesAsync();
+            }
+            return tenant.Id;
+        }
+
         [Authorize(Roles = new[] { RoleNames.Admin })]
-        public async Task<IdentityResult> Update(NewUser user)
+        public async Task<IdentityResult> Update(UpdateUser user)
         {
             var appUser = await userManager.FindByNameAsync(user.UserName);
             appUser = mapper.Map(user, appUser);
